% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cc_defaults}
\alias{cc_defaults}
\title{default of credit card clients}
\format{
A data frame with 30000 observations. I have relabelled variables according to its meaning, and the definition provided in the paper. Reader should examine them beforehand, as some inconsistencies may occur.

\describe{
\item{limit_bal}{Amount of the given credit (NT dollar): it includes both the individual consumer credit and his/her family (supplementary) credit.}
\item{sex}{Gender (1 = male; 2 = female).}
\item{education}{Education (1 = graduate school; 2 = university; 3 = high school; 4 = others).}
\item{marriage}{Marital status (1 = married; 2 = single; 3 = others).}
\item{age}{Age (year).}
\item{pay_sep-pay_abr}{History of past payment. We tracked the past monthly payment records (from April to September, 2005) as follows: pay_sep = the repayment status in September, 2005; pay_aug = the repayment status in August, 2005; ...; pay_apr = the repayment status in April, 2005. The measurement scale for the repayment status is: -2: No consumption; -1: Paid in full; 0: The use of revolving credit; 1 = payment delay for one month; 2 = payment delay for two months; . . .; 8 = payment delay for eight months; 9 = payment delay for nine months and above.}
\item{bill_amt_sep-bill_amt_apr}{Amount of bill statement (NT dollar). bill_amt_sep = amount of bill statement in September, 2005; bill_amt_aug = amount of bill statement in August, 2005; ...; bill_amt_apr = amount of bill statement in April, 2005.}
\item{pay_amt_sep-pay_amt_apr}{Amount of previous payment (NT dollar). pay_amt_sep = amount paid in September, 2005; pay_amt_aug = amount paid in August, 2005; ...;pay_amt_apr = amount paid in April, 2005.}
\item{default}{The target variable, indicating if the customer has defaulted or not (no=0, yes=1).}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/default+of+credit+card+clients#}
}
\usage{
cc_defaults
}
\description{
A dataset of default payments in Taiwan, used in Yeh, I. C., & Lien, C. H. (2009). The comparisons of data mining techniques for the predictive accuracy of probability of default of credit card clients. Expert Systems with Applications, 36(2), 2473-2480.
}
\examples{

cc_defaults

}
\keyword{datasets}
