% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{epub}
\alias{epub}
\title{A sample of epub transactions for rule mining}
\format{
A data frame with 25893 observations and three variables.

\describe{
\item{transaction_id}{The id of the transaction. A transaction appears in as many rows as epub acquired in it.}
\item{time_stamp}{Time stamp of the transaction in POSIXct format.}
\item{book_code}{The code of the epub}
}
}
\source{
Original data provided by Michael Hahsler from ePub-WU at \url{https://epub.wu-wien.ac.at}.
}
\usage{
epub
}
\description{
The epub dataset is a transformation of the Epub transaction matrix of the arules package into a data frame. The dataset contains the download history of documents from the electronic publication platform of the Vienna University of Economics and Business Administration. The data was recorded between Jan 2003 and Dec 2008.
}
\examples{

epub

}
\keyword{datasets}
