% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hlr}
\alias{hlr}
\title{A sample dataset for hierarchical linear regression}
\format{
A data frame with 100 observations and five variables.

\describe{

\item{happiness}{The criterion or dependent variable.}
\item{age}{A control variable.}
\item{gender}{Another control variable, encoded as character with "Male" and "Female" values.}
\item{friends}{A predictor variable.}
\item{pets}{A predictor variable.}
}
}
\source{
Hierarchical linear regression (University of Virginia Library) \url{https://data.library.virginia.edu/hierarchical-linear-regression/}
}
\usage{
hlr
}
\description{
A (probably artificial) dataset to introduce hierarchical linear gression model.
}
\examples{

hrl

}
\keyword{datasets}
