% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hpiUS}
\alias{hpiUS}
\title{house price index for the United States}
\format{
A data frame with 26,877 observations.

\describe{
\item{year}{Year.}
\item{month}{Month.}
\item{state}{US State.}
\item{price_index}{Calculated House Price Index - average price changes in repeat sales or refinancings at state level.}
\item{us_avg}{Calculated House Price Index - averaged at national level.}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2019/2019-02-05}
}
\usage{
hpiUS
}
\description{
The House Price Index (HPI) is a broad measure of the movement of single-family house prices. The HPI is a weighted, repeat-sales index, meaning that it measures average price changes in repeat sales or refinancings on the same properties. This information is obtained by reviewing repeat mortgage transactions on single-family properties whose mortgages have been purchased or securitized by Fannie Mae or Freddie Mac since January 1975.
}
\examples{

hpiUS

}
\keyword{datasets}
