% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ibm_attrition}
\alias{ibm_attrition}
\title{A employee attition dataset}
\format{
An object of class \code{data.frame} with 1470 rows and 35 columns.
}
\source{
\url{https://www.kaggle.com/datasets/pavansubhasht/ibm-hr-analytics-attrition-dataset}
}
\usage{
ibm_attrition
}
\description{
Uncover the factors that lead to employee attrition and explore important questions such as ‘show me a breakdown of distance from home by job role and attrition’ or ‘compare average monthly income by education and attrition’. This is a fictional data set created by IBM data scientists.
}
\examples{

ibm_attrition

}
\keyword{datasets}
