% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{InsuranceCharges}
\alias{InsuranceCharges}
\title{Data for insurance premium prediction}
\format{
A data frame with 1338 observations (rows) and 7 features (columns). The dataset contains four numerical features (age, bmi, children and expenses) and three nominal features (sex, smoker and region) converted into factors with numerical value designated for each level. The variable to predict is charges.
}
\source{
\url{https://www.kaggle.com/noordeen/insurance-premium-prediction}. Original source was  from the Machine Learning course website (Spring 2017) from Professor Eric Suess.
}
\usage{
InsuranceCharges
}
\description{
A dataset to predict the insurance premium (charges).
}
\examples{

InsuranceCharges

}
\keyword{datasets}
