% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mortgageUS}
\alias{mortgageUS}
\title{A collection of indicators of mortgage rates for the United States}
\format{
A data frame with 2,492 observations.

\describe{
\item{date}{Date (in date format).}
\item{fixed_rate_30_yr}{Fixed rate 30 year mortgage (percent).}
\item{fees_and_pts_30_yr}{Fees and percentage points of the loan amount.}
\item{fixed_rate_15_yr}{Fixed rate 15 year mortgage (percent).}
\item{fees_and_pts_15_yr}{Fixed rate 15 year mortgage (percent).}
\item{adjustable_rate_5_1_hybrid}{5-1 Hybrid Adjustable rate mortgage (5 year fixed, then annual adjustable rate).}
\item{fees_and_pts_5_1_hybrid}{Fees and percentage points of the loan amount.}
\item{adjustable_margin_5_1_hybrid}{A fixed amount added to the underlying index to establish the fully indexed rate for an ARM.}
\item{spread_30_yr_fixed_and_5_1_adjustable}{Difference in rate between 30 year fixed and 5-1 adjustable.}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2019/2019-02-05}
}
\usage{
mortgageUS
}
\description{
Mortgage rates for fixed 30, fixed 15, and adjustable 5-1 Hybrids across time. Some of the indices were only introduced in the 1990s and late 2000s, and as such some information can only be pulled from more recent data.
}
\examples{

mortgageUS

}
\keyword{datasets}
