% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{online_retail}
\alias{online_retail}
\title{A dataset of online retail transactions}
\format{
A data frame with 541909 observations and eight variables.

\describe{
\item{InvoiceNo}{Invoice number. Nominal, a 6-digit integral number uniquely assigned to each transaction. If this code starts with letter 'c', it indicates a cancellation.}
\item{StockCode}{Product (item) code. Nominal, a 5-digit integral number uniquely assigned to each distinct product.}
\item{Description}{Product (item) name. Nominal.}
\item{Quantity}{The quantities of each product (item) per transaction. Numeric.}
\item{InvoiceDate}{Invice Date and time. Numeric, the day and time when each transaction was generated. In POSIXct format.}
\item{UnitPrice}{Unit price. Numeric, Product price per unit in sterling.}
\item{CustomerID}{Customer number. Nominal, a 5-digit integral number uniquely assigned to each customer.}
\item{Country}{Country name. Nominal, the name of the country where each customer resides.}
}
}
\source{
UCI Machine Learning Repository \url{https://archive.ics.uci.edu/ml/datasets/online+retail}
}
\usage{
online_retail
}
\description{
A transnational data set which contains all the transactions occurring between 01/12/2010 and 09/12/2011 for a UK-based and registered non-store online retail.The company mainly sells unique all-occasion gifts. Many customers of the company are wholesalers.
}
\examples{

online_retail

}
\keyword{datasets}
