% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{recessionUS}
\alias{recessionUS}
\title{Some recession dates in the US}
\format{
A data frame with 14 observations

\describe{
\item{name}{Name of the recession.}
\item{period_range}{Period from start to finish.}
\item{duration_months}{Duration in months.}
\item{time_since_previous_recession_months}{Time since previous recession in months.}
\item{peak_unemploy_ment}{Peak unemployment and time when it is reached.}
\item{gdp_decline_peak_to_trough}{gdp decline caused by the recession.}
\item{characteristics}{A description of the characteristics of the recession.}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2019/2019-02-05}
}
\usage{
recessionUS
}
\description{
Some recession dates in the US
}
\examples{

recessionUS

}
\keyword{datasets}
