% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{titanic_raw}
\alias{titanic_raw}
\title{A sample of the Titanic dataset for rule mining}
\format{
A data frame with 2201 observations and four variables.

\describe{
\item{Class}{Class where the passenger was travelling (1st, 2nd, 3rd, Crew).}
\item{Sex}{Passenger's gender (Female, Male).}
\item{Age}{Is the passenger an adult or a child? (Adult, Child).}
\item{Survied}{Has the passenger survived? (Yes, No)}
}
}
\source{
\url{http://www.rdatamining.com/data}
}
\usage{
titanic_raw
}
\description{
A sample of the Titanic dataset for rule mining
}
\examples{

titanic_raw

}
\keyword{datasets}
